Move Header container out to it's own reusable component
