+++ WordPress/Classes/Services/PageCoordinator.swift
@@ -13,13 +13,17 @@ class PageCoordinator {
 
     private static func showLayoutPicker(from controller: UIViewController, forBlog blog: Blog, _ completion: @escaping TemplateSelectionCompletion) {
         let storyboard = UIStoryboard(name: "LayoutPickerStoryboard", bundle: Bundle.main)
-        guard let navigationController = storyboard.instantiateInitialViewController() as? UINavigationController,
-            let rootView = navigationController.topViewController as? GutenbergLayoutPickerViewController  else {
+        guard let childViewController = storyboard.instantiateInitialViewController() as? GutenbergLayoutPickerViewController else {
             completion(nil)
             return
         }
-        rootView.completion = completion
-        rootView.blog = blog
+
+        childViewController.completion = completion
+        childViewController.blog = blog
+
+        let container = CollapsableHeaderViewController(childViewController: childViewController)
+        childViewController.headerContentsDelegate = container
+        let navigationController = UINavigationController(rootViewController: container)
 
         if #available(iOS 13.0, *) {
             navigationController.modalPresentationStyle = .pageSheet
