+++ WordPress/Classes/ViewRelated/Gutenberg/Layout Picker/GutenbergLayoutPickerViewController.swift
@@ -15,95 +15,23 @@ class GutenbergLayoutSection {
     }
 }
 
-class GutenbergLayoutPickerViewController: UIViewController {
+class GutenbergLayoutPickerViewController: UIViewController, CollapsableHeaderDataSource, CollapsableHeaderDelegate {
+    let defaultActionTitle = NSLocalizedString("Create Blank Page", comment: "Title for button to make a blank page")
+    let primaryActionTitle = NSLocalizedString("Create Page", comment: "Title for button to make a page with the contents of the selected layout")
+    let secondaryActionTitle = NSLocalizedString("Preview", comment: "Title for button to preview a selected layout")
+    let mainTitle = NSLocalizedString("Choose a Layout", comment: "Title for the screen to pick a template for a page")
+    let prompt = NSLocalizedString("Get started by choosing from a wide variety of pre-made page layouts. Or just start with a blank page.", comment: "Prompt for the screen to pick a template for a page")
+    weak var headerContentsDelegate: CollapsableHeaderContentsDelegate?
 
-    @IBOutlet weak var headerBar: UIView!
-    @IBOutlet weak var headerView: UIView!
-    @IBOutlet weak var closeButton: UIButton!
-    @IBOutlet weak var titleView: UILabel!
-    @IBOutlet weak var largeTitleView: UILabel!
-    @IBOutlet weak var promptView: UILabel!
-    @IBOutlet weak var filterBar: GutenbergLayoutFilterBar!
     @IBOutlet weak var tableView: UITableView!
-    @IBOutlet weak var footerView: UIView!
-    @IBOutlet weak var createBlankPageBtn: UIButton!
-    @IBOutlet weak var previewBtn: UIButton!
-    @IBOutlet weak var createPageBtn: UIButton!
-    @IBOutlet weak var selectedStateButtonsContainer: UIView!
-
-    /// This  is used as a means to adapt to different text sizes to force the desired layout and then active `headerHeightConstraint`
-    /// when scrolling begins to allow pushing the non static items out of the scrollable area.
-    @IBOutlet weak var initialHeaderTopConstraint: NSLayoutConstraint!
-    @IBOutlet weak var headerHeightConstraint: NSLayoutConstraint!
-    @IBOutlet weak var titleToSubtitleSpacing: NSLayoutConstraint!
-    @IBOutlet weak var subtitleToCategoryBarSpacing: NSLayoutConstraint!
-    @IBOutlet weak var minHeaderBottomSpacing: NSLayoutConstraint!
-    @IBOutlet weak var maxHeaderBottomSpacing: NSLayoutConstraint!
-    @IBOutlet var visualEffects: [UIVisualEffectView]! {
-        didSet {
-            if #available(iOS 13.0, *) {
-                visualEffects.forEach { (visualEffect) in
-                    visualEffect.effect = UIBlurEffect.init(style: .systemChromeMaterial)
-                }
-            }
-        }
-    }
-
-    private var shouldUseCompactLayout: Bool {
-        return traitCollection.verticalSizeClass == .compact
-    }
-
-    private var _maxHeaderHeight: CGFloat = 0
-    private var maxHeaderHeight: CGFloat {
-        if shouldUseCompactLayout {
-            return minHeaderHeight
-        } else {
-            return _maxHeaderHeight
-        }
-    }
-
-    private var _midHeaderHeight: CGFloat = 0
-    private var midHeaderHeight: CGFloat {
-        if shouldUseCompactLayout {
-            return minHeaderHeight
-        } else {
-            return _midHeaderHeight
-        }
-    }
-    private var minHeaderHeight: CGFloat = 0
-
-    private var titleIsHidden: Bool = true {
-        didSet {
-            if oldValue != titleIsHidden {
-                titleView.isHidden = false
-                let alpha: CGFloat = titleIsHidden ? 0 : 1
-                UIView.animate(withDuration: 0.4, delay: 0, options: .transitionCrossDissolve, animations: {
-                    self.titleView.alpha = alpha
-                }) { (_) in
-                    self.titleView.isHidden = self.titleIsHidden
-                }
-            }
-        }
-    }
-
-    var accentColor: UIColor {
-        if #available(iOS 13.0, *) {
-            return UIColor { (traitCollection: UITraitCollection) -> UIColor in
-                if traitCollection.userInterfaceStyle == .dark {
-                    return UIColor.muriel(color: .accent, .shade40)
-                } else {
-                    return UIColor.muriel(color: .accent, .shade50)
-                }
-            }
-        } else {
-            return UIColor.muriel(color: .accent, .shade50)
-        }
+    var scrollView: UIScrollView {
+        return tableView
     }
 
     private var selectedLayout: IndexPath? = nil {
         didSet {
             if !(oldValue != nil && selectedLayout != nil) {
-                layoutSelected(selectedLayout != nil)
+                headerContentsDelegate?.itemSelectionChanged(selectedLayout != nil)
             }
         }
     }
@@ -118,8 +46,7 @@ class GutenbergLayoutPickerViewController: UIViewController {
 
     private var isLoading: Bool = true {
         didSet {
-            filterBar.shouldShowGhostContent = isLoading
-            filterBar.allowsMultipleSelection = !isLoading
+            headerContentsDelegate?.loadingStateChanged(isLoading)
             if isLoading {
                 tableView.startGhostAnimation(style: GhostCellStyle.muriel)
             } else {
@@ -127,45 +54,16 @@ class GutenbergLayoutPickerViewController: UIViewController {
             }
 
             tableView.reloadData()
-            filterBar.reloadData()
         }
     }
 
     var completion: PageCoordinator.TemplateSelectionCompletion? = nil
     var blog: Blog? = nil
 
-    private func setStaticText() {
-        closeButton.accessibilityLabel = NSLocalizedString("Close", comment: "Dismisses the current screen")
-
-        let translatedTitle = NSLocalizedString("Choose a Layout", comment: "Title for the screen to pick a template for a page")
-        titleView.text = translatedTitle
-        largeTitleView.text = translatedTitle
-
-        promptView.text = NSLocalizedString("Get started by choosing from a wide variety of pre-made page layouts. Or just start with a blank page.", comment: "Prompt for the screen to pick a template for a page")
-        createBlankPageBtn.setTitle(NSLocalizedString("Create Blank Page", comment: "Title for button to make a blank page"), for: .normal)
-        previewBtn.setTitle(NSLocalizedString("Preview", comment: "Title for button to preview a selected layout"), for: .normal)
-        createPageBtn.setTitle(NSLocalizedString("Create Page", comment: "Title for button to make a page with the contents of the selected layout"), for: .normal)
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
         tableView.register(LayoutPickerSectionTableViewCell.nib, forCellReuseIdentifier: LayoutPickerSectionTableViewCell.cellReuseIdentifier)
         fetchLayouts()
-        filterBar.filterDelegate = self
-        setStaticText()
-        closeButton.setImage(UIImage.gridicon(.crossSmall), for: .normal)
-        styleButtons()
-        layoutHeader()
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        navigationController?.setNavigationBarHidden(true, animated: true)
-        super.viewWillAppear(animated)
-    }
-
-    override func viewDidDisappear(_ animated: Bool) {
-        navigationController?.isNavigationBarHidden = false
-        super.viewDidDisappear(animated)
     }
 
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -186,37 +84,11 @@ class GutenbergLayoutPickerViewController: UIViewController {
     override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         super.traitCollectionDidChange(previousTraitCollection)
 
-        if #available(iOS 13.0, *) {
-            if traitCollection.hasDifferentColorAppearance(comparedTo: previousTraitCollection) {
-                styleButtons()
-            }
-        }
-
         if let previousTraitCollection = previousTraitCollection, traitCollection.verticalSizeClass != previousTraitCollection.verticalSizeClass {
-            layoutTableViewHeader()
             if let visibleRow = tableView.indexPathsForVisibleRows?.first {
                 tableView.scrollToRow(at: visibleRow, at: .top, animated: true)
             }
         }
-    }
-
-    @IBAction func closeModal(_ sender: Any) {
-        dismiss(animated: true, completion: nil)
-    }
-
-    @IBAction func createBlankPageTapped(_ sender: Any) {
-        createPage(layout: nil)
-    }
-
-    @IBAction func createPageTapped(_ sender: Any) {
-        guard let sectionIndex = selectedLayout?.section, let position = selectedLayout?.item else {
-            createPage(layout: nil)
-            return
-        }
-
-        let layout = sections[sectionIndex].layouts[position]
-        LayoutPickerAnalyticsEvent.templateApplied(slug: layout.slug)
-        createPage(layout: layout)
     }
 
     private func createPage(layout: PageTemplateLayout?) {
@@ -227,101 +99,6 @@ class GutenbergLayoutPickerViewController: UIViewController {
 
         dismiss(animated: true) {
             completion(layout)
-        }
-    }
-
-    private func styleButtons() {
-        let seperator: UIColor
-        if #available(iOS 13.0, *) {
-            seperator = .separator
-        } else {
-            seperator = UIColor.muriel(color: .divider)
-        }
-
-        [createBlankPageBtn, previewBtn].forEach { (button) in
-            button?.titleLabel?.font = WPStyleGuide.fontForTextStyle(.body, fontWeight: .medium)
-            button?.layer.borderColor = seperator.cgColor
-            button?.layer.borderWidth = 1
-            button?.layer.cornerRadius = 8
-        }
-
-        createPageBtn.titleLabel?.font = WPStyleGuide.fontForTextStyle(.body, fontWeight: .medium)
-        createPageBtn.backgroundColor = accentColor
-        createPageBtn.layer.cornerRadius = 8
-
-        if #available(iOS 13.0, *) {
-            closeButton.backgroundColor = UIColor { (traitCollection: UITraitCollection) -> UIColor in
-                if traitCollection.userInterfaceStyle == .dark {
-                    return UIColor.systemFill
-                } else {
-                    return UIColor.quaternarySystemFill
-                }
-            }
-        }
-    }
-
-    private func layoutTableViewHeader() {
-        let topInset = maxHeaderHeight + headerBar.frame.height
-        tableView.tableHeaderView = UIView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: topInset))
-        updateTableViewFooter()
-    }
-
-    /*
-     * Calculates the needed space for the footer to allow the header to still collapse but also to prevent unneeded space
-     * at the bottome of the tableView when multiple cells are rendered.
-     */
-    private func updateTableViewFooter() {
-        let estimatedRowHeight: CGFloat = LayoutPickerSectionTableViewCell.estimatedCellHeight
-        let minimumFooterSize = footerView.frame.size.height + (UIApplication.shared.keyWindow?.safeAreaInsets.bottom ?? 0)
-        let rowCount = CGFloat(max((filteredSections ?? sections).count, 1))
-
-        /// The needed distance to fill the rest of the screen to allow the header to still collapse when scrolling (or to maintain a collapsed header if it was already collapsed when selecting a filter)
-        let distanceToBottom = tableView.frame.height - headerBar.frame.height - minHeaderHeight - (estimatedRowHeight * rowCount)
-        let newHeight: CGFloat = max(minimumFooterSize, distanceToBottom)
-
-        tableView.tableFooterView = UIView(frame: CGRect(x: 0, y: 0, width: tableView.frame.width, height: newHeight))
-        tableView.tableFooterView?.isGhostableDisabled = true
-    }
-
-    private func calculateHeaderSnapPoints() {
-        minHeaderHeight = filterBar.frame.height + minHeaderBottomSpacing.constant
-        _midHeaderHeight = titleToSubtitleSpacing.constant + promptView.frame.height + subtitleToCategoryBarSpacing.constant + filterBar.frame.height + maxHeaderBottomSpacing.constant
-        _maxHeaderHeight = largeTitleView.frame.height + _midHeaderHeight
-    }
-
-    private func layoutHeader() {
-        largeTitleView.font = WPStyleGuide.serifFontForTextStyle(UIFont.TextStyle.largeTitle, fontWeight: .semibold)
-        titleView.font = WPStyleGuide.serifFontForTextStyle(UIFont.TextStyle.largeTitle, fontWeight: .semibold).withSize(17)
-
-        headerView.setNeedsLayout()
-        headerView.layoutIfNeeded()
-        footerView.setNeedsLayout()
-        footerView.layoutIfNeeded()
-
-        calculateHeaderSnapPoints()
-        layoutTableViewHeader()
-
-        tableView.tableHeaderView?.backgroundColor = .clear
-        tableView.tableFooterView?.backgroundColor = .clear
-    }
-
-    private func layoutSelected(_ isSelected: Bool) {
-        createBlankPageBtn.isHidden = false
-        selectedStateButtonsContainer.isHidden = false
-
-        createBlankPageBtn.alpha = isSelected ? 1 : 0
-        selectedStateButtonsContainer.alpha = isSelected ? 0 : 1
-
-        let alpha: CGFloat = isSelected ? 0 : 1
-        let selectedStateContainerAlpha: CGFloat = isSelected ? 1 : 0
-
-
-        UIView.animate(withDuration: LayoutPickerCollectionViewCell.selectionAnimationSpeed, delay: 0, options: .transitionCrossDissolve, animations: {
-            self.createBlankPageBtn.alpha = alpha
-            self.selectedStateButtonsContainer.alpha = selectedStateContainerAlpha
-        }) { (_) in
-            self.createBlankPageBtn.isHidden = isSelected
-            self.selectedStateButtonsContainer.isHidden = !isSelected
         }
     }
 
@@ -337,69 +114,24 @@ class GutenbergLayoutPickerViewController: UIViewController {
             return GutenbergLayoutSection(category)
         }) ?? []
     }
-}
 
-extension GutenbergLayoutPickerViewController: UITableViewDelegate {
+    func defaultActionSelected() {
+        createPage(layout: nil)
+    }
 
-    func scrollViewDidScroll(_ scrollView: UIScrollView) {
-        guard !shouldUseCompactLayout else {
-            titleIsHidden = false
+    func primaryActionSelected() {
+        guard let sectionIndex = selectedLayout?.section, let position = selectedLayout?.item else {
+            createPage(layout: nil)
             return
         }
 
-        if !headerHeightConstraint.isActive {
-            initialHeaderTopConstraint.isActive = false
-            headerHeightConstraint.isActive = true
-        }
-
-        let scrollOffset = scrollView.contentOffset.y
-        let newHeaderViewHeight = maxHeaderHeight - scrollOffset
-
-        if newHeaderViewHeight < minHeaderHeight {
-            headerHeightConstraint.constant = minHeaderHeight
-        } else {
-            headerHeightConstraint.constant = newHeaderViewHeight
-        }
-
-        titleIsHidden = largeTitleView.frame.maxY > 0
+        let layout = sections[sectionIndex].layouts[position]
+        LayoutPickerAnalyticsEvent.templateApplied(slug: layout.slug)
+        createPage(layout: layout)
     }
 
-    func scrollViewDidEndDecelerating(_ scrollView: UIScrollView) {
-        snapToHeight(scrollView)
-    }
-
-    func scrollViewDidEndDragging(_ scrollView: UIScrollView, willDecelerate decelerate: Bool) {
-        if !decelerate {
-            snapToHeight(scrollView)
-        }
-    }
-
-    private func snapToHeight(_ scrollView: UIScrollView) {
-        guard !shouldUseCompactLayout else { return }
-
-        if largeTitleView.frame.midY > 0 {
-            snapToHeight(scrollView, height: maxHeaderHeight)
-        } else if promptView.frame.midY > 0 {
-            snapToHeight(scrollView, height: midHeaderHeight)
-        } else if headerHeightConstraint.constant != minHeaderHeight {
-            snapToHeight(scrollView, height: minHeaderHeight)
-        }
-    }
-
-    private func snapToHeight(_ scrollView: UIScrollView, height: CGFloat) {
-        scrollView.contentOffset.y = maxHeaderHeight - height
-        headerHeightConstraint.constant = height
-        titleIsHidden = (height >= maxHeaderHeight) && !shouldUseCompactLayout
-        UIView.animate(withDuration: 0.4, delay: 0, usingSpringWithDamping: 0.7, initialSpringVelocity: 0.5, options: .curveEaseInOut, animations: {
-            self.headerView.setNeedsLayout()
-            self.headerView.layoutIfNeeded()
-        }, completion: nil)
-    }
-
-    private func containsSelectedLayout(_ selectedIndexPath: IndexPath, atIndexPath indexPath: IndexPath) -> Bool {
-        let rowSection = (filteredSections ?? sections)[indexPath.row]
-        let sectionSlug = sections[selectedIndexPath.section].section.slug
-        return (sectionSlug == rowSection.section.slug)
+    func secondaryActionSelected() {
+        performSegue(withIdentifier: "preview", sender: nil)
     }
 }
 
@@ -430,6 +162,19 @@ extension GutenbergLayoutPickerViewController: UITableViewDataSource {
         }
 
         return cell
+    }
+
+    private func containsSelectedLayout(_ selectedIndexPath: IndexPath, atIndexPath indexPath: IndexPath) -> Bool {
+        let rowSection = (filteredSections ?? sections)[indexPath.row]
+        let sectionSlug = sections[selectedIndexPath.section].section.slug
+        return (sectionSlug == rowSection.section.slug)
+    }
+
+    func estimatedContentSize() -> CGSize {
+        let rowCount = CGFloat(max((filteredSections ?? sections).count, 1))
+        let estimatedRowHeight: CGFloat = LayoutPickerSectionTableViewCell.estimatedCellHeight
+        let estimatedHeight = (estimatedRowHeight * rowCount)
+        return CGSize(width: tableView.contentSize.width, height: estimatedHeight)
     }
 }
 
@@ -466,13 +211,13 @@ extension GutenbergLayoutPickerViewController: LayoutPickerSectionTableViewCellD
     }
 }
 
-extension GutenbergLayoutPickerViewController: FilterBarDelegate {
+extension GutenbergLayoutPickerViewController: CollapsableHeaderFilterBarDelegate {
     func numberOfFilters() -> Int {
         return sections.count
     }
 
-    func filter(forIndex index: Int) -> GutenbergLayoutSection {
-        return sections[index]
+    func filter(forIndex index: Int) -> CollabsableHeaderFilterOption {
+        return sections[index].section
     }
 
     func didSelectFilter(withIndex selectedIndex: IndexPath, withSelectedIndexes selectedIndexes: [IndexPath]) {
@@ -485,7 +230,7 @@ extension GutenbergLayoutPickerViewController: FilterBarDelegate {
 
         filteredSections = [sections[selectedIndex.item]]
         tableView.performBatchUpdates({
-            updateTableViewFooter()
+            headerContentsDelegate?.contentSizeWillChange()
             tableView.deleteRows(at: rowsToRemove, with: .fade)
         })
     }
@@ -501,7 +246,7 @@ extension GutenbergLayoutPickerViewController: FilterBarDelegate {
 
         tableView.performBatchUpdates({
             if selectedIndexes.count == 2 {
-                updateTableViewFooter()
+                headerContentsDelegate?.contentSizeWillChange()
             }
             tableView.reloadSections([0], with: .automatic)
         })
@@ -515,7 +260,7 @@ extension GutenbergLayoutPickerViewController: FilterBarDelegate {
 
         filteredSections = nil
         tableView.performBatchUpdates({
-            updateTableViewFooter()
+            headerContentsDelegate?.contentSizeWillChange()
             tableView.reloadSections([0], with: .fade)
         })
     }
@@ -536,7 +281,7 @@ extension GutenbergLayoutPickerViewController: FilterBarDelegate {
 
         guard let rowToRemove = row else { return }
         tableView.performBatchUpdates({
-            updateTableViewFooter()
+            headerContentsDelegate?.contentSizeWillChange()
             tableView.deleteRows(at: [rowToRemove], with: .fade)
         })
     }
@@ -548,6 +293,5 @@ extension GutenbergLayoutPickerViewController: NSFetchedResultsControllerDelegat
         sections = makeSectionData(with: resultsController)
         isLoading = resultsController.isEmpty()
         tableView.reloadData()
-        filterBar.reloadData()
     }
 }
